+++ tests/test_regression.py
@@ -6063,6 +6063,20 @@ def test_2723():
     dot("png", input)
 
 
+def test_2727():
+    """
+    the label “<>” should be accepted
+    https://gitlab.com/graphviz/graphviz/-/issues/2727
+    """
+
+    # locate our associated test case in this directory
+    input = Path(__file__).parent / "2727.dot"
+    assert input.exists(), "unexpectedly missing test case"
+
+    # process this
+    dot("svg", input)
+
+
 @pytest.mark.parametrize("package", ("Tcldot", "Tclpathplan"))
 @pytest.mark.skipif(shutil.which("tclsh") is None, reason="tclsh not available")
 @pytest.mark.xfail(